requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
anyio==3.7.1
astunparse==1.6.3
boto3==1.37.38
botocore==1.37.38
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==7.1.2
cryptography==44.0.2
dill==0.3.6
dnslib==0.9.26
dnspython==2.6.1
ecdsa==0.19.1
exceptiongroup==1.2.2
fastapi==0.115.12
ffmpeg-python==0.2.0
flatbuffers==25.2.10
future==1.0.0
gast==0.4.0
google-auth==2.39.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.70.0
h11==0.12.0
h2==4.1.0
h5py==3.11.0
hpack==4.0.0
httpcore==0.13.7
httpx==0.19.0
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.5.0
jax==0.4.13
jmespath==1.0.1
keras==2.12.0
libclang==18.1.1
localstack==3.8.1
localstack-core==3.8.1
localstack-ext==3.8.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.2.0
norbert==0.2.1
numpy==1.24.3
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==25.0
pandas==1.5.3
plux==1.12.1
protobuf==4.25.7
psutil==7.0.0
pyaes==1.6.1
pyasn1==0.4.8
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
pyotp==2.9.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rfc3986==1.5.0
rich==14.0.0
rsa==4.9.1
s3transfer==0.11.5
scipy==1.10.1
semver==3.0.4
six==1.17.0
sniffio==1.3.1
spleeter==2.4.2
starlette==0.44.0
tabulate==0.9.0
tailer==0.4.1
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.1
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.4.0
tomli==2.2.1
typer==0.3.2
typing_extensions==4.13.2
urllib3==1.26.20
uvicorn==0.33.0
Werkzeug==3.0.6
wrapt==1.14.1
zipp==3.20.2
